move seqs base types into seqs.go; better appendStringExchange implementation
